refactor: sandbox page header
